git user data from google relates #13

+++ server/controller/index.js
@@ -1,5 +1,15 @@
-const {register,getAllUsers,login,logout,deleteAllUsers}=require('./user');
+const {register,getAllUsers,login,logout,deleteAllUsers,googleAuth}=require('./user');
 const {getAllBlogs,addNewBlog,deleteAllBlogs}=require('./blog/blogs')
 
 
-module.exports={register,getAllUsers,login,logout,getAllBlogs,addNewBlog,deleteAllBlogs,deleteAllUsers}
+module.exports={
+    register,
+    getAllUsers,
+    login,
+    logout,
+    getAllBlogs,
+    addNewBlog,
+    deleteAllBlogs,
+    deleteAllUsers,
+    googleAuth
+}
